some stuff

+++ Assets/_Project/Scripts/SO/GhostieStats.cs
@@ -8,7 +8,7 @@ public class GhostieStats : ScriptableObject
     [field: SerializeField] public float Speed { get; private set; } = 2;
     [field: SerializeField] public float RotationSpeed { get; private set; } = 180;
     [field: SerializeField] public float ChaseSpeed { get; private set; } = 3;
-    [field: SerializeField] public float Damage { get; private set; } = 1;
+    [field: SerializeField] public int Damage { get; private set; } = 1;
     [field: SerializeField] public int Health { get; private set; } = 3;
     [field: SerializeField] public float SightRange { get; private set; } = 8;
     [field: SerializeField] public float ForgetRange { get; private set; } = 14;
